GCI-1428: Reduce test code duplication.
